09-08 곽성빈 아이템 줍기
 아이템 줍기 + 아이템 부가 스텟 추가

+++ Assets/Scrip/Item/Item_Data.cs
@@ -7,7 +7,7 @@ public class Item_Data : MonoBehaviour
     [Header("Item Information")]
     public bool Is_Twohand;
     public int Damage;
-    public int Add_Hp;
+    public float Add_Hp;
     public Equipment_Type Item_Type;
 
     public float Atk_Speed;
